verify-ppb/pharmacists/requirements-dev.txt
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Development
ipython==8.18.1
